proj/crypto_test/requirements.txt
selenium
Pyyaml
pytest
allure-pytest
appium-python-client
requests